requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
python-Levenshtein>=0.25.0
pyautogui>=0.9.54
pyperclip>=1.9.0
psutil>=5.9.8
requests>=2.31.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.2
